+++ app/Http/Controllers/AttributeGroupController.php
@@ -30,8 +30,14 @@ class AttributeGroupController extends Controller {
         $this->rightObj = new Right();
     }
 
-    public function index(){        
+    public function index(){
+      $rightId=121;  
+      $currentChannelId = $this->rightObj->getCurrnetChannelId($rightId);
+      $channels = $this->rightObj->getAllowedChannels($rightId);  
+      if(!$this->rightObj->checkRights($currentChannelId, $rightId))
+            return redirect('/dashboard');
        $groups=  AttributeGroup::join('product_types','attribute_groups.product_type_id','=','product_types.id')
+               ->where('channel_id',$currentChannelId)
                ->select('attribute_groups.*','product_types.name as type_name')
                ->orderBy('created_at');
        if (isset($_GET['keyword'])) {
@@ -41,7 +47,7 @@ class AttributeGroupController extends Controller {
            }
        }
        $groups=$groups->paginate(config('constants.recordperpage'));
-       return view('attributegroup.attribute_groups',compact('groups'));
+       return view('attributegroup.attribute_groups',compact('groups','channels','currentChannelId'));
     }
 
     /**
@@ -51,13 +57,27 @@ class AttributeGroupController extends Controller {
      * @return Response
      */
     public function create(){
-      $productTypes= ProductType::orderBy('name','desc')->lists('name','id')->toArray();  
+      $rightId=122;  
+      $currentChannelId = $this->rightObj->getCurrnetChannelId($rightId);
+      $channels = $this->rightObj->getAllowedChannels($rightId);
+      if(!$this->rightObj->checkRights($currentChannelId, $rightId))
+            return redirect('/dashboard');  
+      $allowedChannel=array();
+      foreach($channels as $channel){
+          $allowedChannel[]=$channel->channel_id;
+      }
+      
+      $productTypes= ProductType::whereIn('channel_id',$allowedChannel)->orderBy('name','desc')->lists('name','id')->toArray();  
       return view('attributegroup.create',compact('productTypes'));
     }
 
     public function store(Request $request) {
-      
-          $validation = $this->validate($request,[
+        $rightId=122;   
+        $currentPtype=ProductType::find($request->product_type);
+        if(!$this->rightObj->checkRights($currentPtype->channel_id, $rightId))
+              return redirect('/dashboard');
+        
+        $validation = $this->validate($request,[
             'product_type' => 'required',  
             'group_name' => 'required',
         ]);
@@ -74,9 +94,19 @@ class AttributeGroupController extends Controller {
 
   
     public function edit($id){       
-        $rightId = 10;
-        $productTypes= ProductType::orderBy('name','desc')->lists('name','id')->toArray();
-        $group= AttributeGroup::find($id); // Brand::find($id);  
+        $rightId=122;  
+        $group= AttributeGroup::find($id); // Brand::find($id); 
+        $currentPtype=ProductType::find($group->product_type_id);
+        $currentChannelId = $currentPtype->channel_id;
+        $channels = $this->rightObj->getAllowedChannels($rightId);
+        if(!$this->rightObj->checkRights($currentChannelId, $rightId))
+              return redirect('/dashboard');  
+        $allowedChannel=array();
+        foreach($channels as $channel){
+            $allowedChannel[]=$channel->channel_id;
+        }
+        $productTypes= ProductType::whereIn('channel_id',$allowedChannel)->orderBy('name','desc')->lists('name','id')->toArray();
+         
         return view('attributegroup.edit',compact('group','productTypes'));
     }
 
@@ -88,12 +118,16 @@ class AttributeGroupController extends Controller {
      * @return Response
      */
     public function update(Request $request,$id) {
-          
-        $validation = $this->validate($request,[
+        $rightId=122;   
+        $currentPtype=ProductType::find($request->product_type);
+        if(!$this->rightObj->checkRights($currentPtype->channel_id, $rightId))
+              return redirect('/dashboard');
+        
+         $validation = $this->validate($request,[
             'product_type' => 'required',  
             'group_name' => 'required',
         ]);
-        
+         
         $group = AttributeGroup::find($id);
         $group->name = trim($request->group_name);
         $group->product_type_id = $request->product_type;
@@ -118,7 +152,11 @@ class AttributeGroupController extends Controller {
   
     
     public function destroy(Request $request){
-        //dd($request);
+        $rightId=122;
+        $currentAGroup=AttributeGroup::find($request->checkItem[0]);
+        $currentPtype=ProductType::find($currentAGroup->product_type_id);
+        if(!$this->rightObj->checkRights( $currentPtype->channel_id, $rightId))
+              return redirect('/dashboard');
         AttributeGroup::whereIn('id',$request->checkItem)->delete();
         Session::flash('message', 'Attribute groups deleted successfully.');
         return Redirect::to('attribute-groups/');
